requirements.txt
PyQt5==5.11.2
pyserial==3.4